Add ::check pseudo element
This pseudo element is added to allow <option> elements inside
<select> elements to have a standardized and customizable
checkmark icon.

Spec definition:
```
Option checkmark icon: the '::check' pseudo-element

The '::check' pseudo-element is only generated when the originating
element is a option element which has an ancestor select element
with base appearance.

'::check' is a tree-abiding pseudo-element. It accepts all
properties.
It inherits from its originating element.

'::check' generates a box as if it was an child of its originating
element, preceding any boxes generated by the '::before'
pseudo-element, with content as specified by 'content'.
```

Defined in:
https://github.com/w3c/csswg-drafts/pull/10986

Bug: 369319576
Change-Id: Ie5058d66f90644f9732bc619a3f587b6203d1cf9