[Gap Decorations] Introduce ColumnRuleColor on ComputedStyle
This CL introduces new data structures to extend the ColumnRuleColor
property for the gap decorations feature. The following classes are
introduced:

StyleColorRepeater: Holds color repeater values.
GapColorData: Represents either a single StyleColor or a StyleColorRepeater.
GapColorDataList: Contains a list of GapColorData.

The column-rule-color type is updated to be GapColorDataList, and
relevant call sites have been modified accordingly.

Additionally, this CL enhances the gap decorations VirtualTestSuites
to cover more cases with the feature flag enabled, using existing tests.
It also adds new tests related to animation and serialization to the
TestExpectations file, which are currently failing with the feature
flag enabled.

The code is exercised through the gap-decorations-color-computed.html
test.

Bug: 357648037
Change-Id: Iebfe4c4610a6f5375a913a64e7d845ba2b2b7e65